--- hooks/admin/recruit/useInfiniteRecruitList.ts
@@ -1,41 +1,0 @@
-import { useEffect, useRef } from 'react';
-import { recruitListData } from 'types/admin/adminRecruitmentsTypes';
-import { instance } from 'utils/axios';
-import { InfiniteScroll } from 'utils/infinityScroll';
-
-const fetchRecruitList = (page: number) => {
-  return instance
-    .get(`/admin/recruitments?page=${page}&size=${8}`)
-    .then((res) => {
-      return res.data;
-    });
-};
-
-const useInfiniteRecruitList = () => {
-  const { data, isLoading, isError, fetchNextPage, hasNextPage } =
-    InfiniteScroll<recruitListData>(['recruitList'], fetchRecruitList, 'JY09');
-
-  // 감지할 대상
-  const targetRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver((entries) => {
-      if (entries[0].isIntersecting && hasNextPage) {
-        fetchNextPage();
-      }
-    });
-    const currentTarget = targetRef.current;
-    if (currentTarget && hasNextPage) {
-      observer.observe(currentTarget);
-      return;
-    }
-    if (currentTarget && !hasNextPage) {
-      observer.unobserve(currentTarget);
-    }
-    console.log(data);
-  }, [fetchNextPage, hasNextPage]);
-
-  return { data, isLoading, isError, targetRef };
-};
-
-export default useInfiniteRecruitList;
